python-app/requirements.txt
flask==2.3.3
requests==2.28.2
numpy==1.26.4
pandas==2.1.4
sqlalchemy==1.4.54
Pillow==12.1.1
PyYAML==6.0.1
cryptography==46.0.5
